requirements.txt
python-telegram-bot==21.0.1
aiosqlite==0.20.0
python-dotenv==1.0.1
pytz==2024.1
gspread==6.1.2
APScheduler==3.10.4
fpdf2==2.7.9
